+++ app/src/main/java/starships/gameObjects/Bullet.java
@@ -4,7 +4,7 @@ import starships.*;
 
 public class Bullet extends GameObject {
 
-    public Bullet(String id, ObjectStyle style, ObjectSize size, Vector position, Vector direction, int speed, int rotationDegrees, int damage, boolean isVisible) {
+    public Bullet(String id, ObjectStyle style, ObjectSize size, Vector position, Vector direction, double speed, double rotationDegrees, double damage, boolean isVisible) {
         super(id, ObjectType.BULLET, style, ObjectShape.RECTANGULAR, size, position, direction, speed, rotationDegrees, damage, 1, isVisible);
     }
